release notes hook
